refactor: move interfaces and type aliases to types.ts + switch to more accurate type-checks with nodejs' util
