+++ src/lib.ts
@@ -4,4 +4,3 @@ export * from './info';
 export * from './command';
 export { spawn } from './process';
 export * from './probe';
-export { FFmpegError, FFprobeError } from './errors';
