+++ src/types.ts
@@ -1,5 +1,4 @@
 import { ChildProcess, SpawnOptionsWithoutStdio } from 'child_process';
-import { ProbeOptions, ProbeResult } from './probe';
 import {
   AudioCodec,
   AudioDecoder,
@@ -33,7 +32,7 @@ export enum LogLevel {
 export type InputSource = string | Uint8Array | AsyncIterable<Uint8Array>;
 /** @public */
 export type OutputDestination = string | NodeJS.WritableStream;
-/** @alpha */
+/** @public */
 export type ConcatSource = InputSource
   | {
     file?: InputSource;
@@ -59,10 +58,8 @@ export interface FFmpegCommand {
    */
   input(source: InputSource): FFmpegInput;
   /**
-   * **UNSTABLE:** New API, see https://github.com/FedericoCarboni/eloquent-ffmpeg/issues/2
-   *
-   * Concatenate media files using the `concat` demuxer.
-   *
+   * Concatenate media files using the `concat` demuxer. This can also be used with stream copy, so
+   * it is much faster than the `concat` filter for most applications.
    * @param sources - The input sources to be concatenated, they can be in different formats but
    * they must have the same streams, codecs, timebases, etc...
    * {@link https://ffmpeg.org/ffmpeg-formats.html#concat-1}
@@ -75,7 +72,6 @@ export interface FFmpegCommand {
    * const process = await cmd.spawn();
    * await process.complete();
    * ```
-   * @alpha
    */
   concat(sources: ConcatSource[], options?: ConcatOptions): FFmpegInput;
   /**
@@ -118,7 +114,7 @@ export interface FFmpegCommand {
   getArgs(): string[];
 }
 
-/** @alpha */
+/** @public */
 export interface ConcatOptions {
   safe?: boolean;
   protocols?: string[];
@@ -201,17 +197,10 @@ export interface FFmpegOptions {
 /** @public */
 export interface FFmpegInput {
   /**
-   * **UNSTABLE**: Breaking changes are being considered, implementation details can change without
-   * notice.
-   *
    * Get information about the input, this is especially helpful when working
    * with streams. If the source is a stream `options.probeSize` number of bytes
    * will be read and passed to ffprobe; those bytes will be kept in memory
    * until the input is used in conversion.
-   *
-   * **Note:** This is not recommended for `concat()` inputs, because it may not
-   * have effect you may expect. When using `concat()` inputs with streams, the
-   * streams will be consumed.
    *
    * @param options -
    * @example
@@ -488,4 +477,276 @@ export interface FFmpegProcess {
   pause(): boolean;
   /** Resumes the conversion, returns `true` if the operation succeeds or `false` if it fails. */
   resume(): boolean;
+}
+
+/* eslint-disable camelcase */
+
+/** @alpha */
+export interface RawProbeResult {
+  format: RawProbeFormat;
+  streams: RawProbeStream[];
+  chapters: RawProbeChapter[];
+  error?: RawProbeError;
+}
+
+/** @alpha */
+export interface RawProbeError {
+  code: number;
+  string: string;
+}
+
+// https://github.com/FFmpeg/FFmpeg/blob/9d8f9b2e4094ae6b07a9f23ae044b802722b3b4e/fftools/ffprobe.c#L2807
+/** @alpha */
+export interface RawProbeFormat {
+  [key: string]: any;
+
+  filename?: string;
+
+  nb_streams: number;
+  nb_programs: number;
+
+  format_name?: string;
+  format_long_name?: string;
+
+  start_time: string;
+  duration: string;
+  size?: string;
+  bit_rate?: string;
+
+  probe_score: number;
+
+  tags?: Record<string, string>;
+}
+
+// https://github.com/FFmpeg/FFmpeg/blob/9d8f9b2e4094ae6b07a9f23ae044b802722b3b4e/fftools/ffprobe.c#L2485
+/** @alpha */
+export type RawProbeStream = {
+  [key: string]: any;
+
+  index: number;
+  codec_name?: string;
+  codec_long_name?: string;
+  profile?: string;
+  codec_type: 'video';
+  codec_time_base?: string;
+  codec_tag_string: string;
+  codec_tag: string;
+
+  id?: string;
+  r_frame_rate?: string;
+  avg_frame_rate?: string;
+  time_base?: string;
+  start_pts?: string;
+  start_time?: string;
+  duration_ts?: string;
+  duration?: string;
+  bit_rate?: string;
+  max_bit_rate?: string;
+  bits_per_raw_sample?: number;
+  nb_frames?: number;
+  nb_read_frames?: number;
+  nb_read_packets?: number;
+  disposition?: RawProbeDisposition;
+
+  width: number;
+  height: number;
+  coded_width?: number;
+  coded_height?: number;
+  closed_captions?: number;
+  has_b_frames: number;
+  sample_aspect_ratio?: string;
+  display_aspect_ratio?: string;
+  pix_fmt?: string;
+  level: number;
+  color_range?: string;
+  color_space?: string;
+  color_primaries?: string;
+  chroma_location?: string;
+  field_order?: string;
+  timecode?: string;
+  refs?: string;
+  tags?: Record<string, string>;
+} | {
+  [key: string]: any;
+
+  index: number;
+  codec_name?: string;
+  codec_long_name?: string;
+  profile?: string;
+  codec_type: 'audio';
+  codec_time_base?: string;
+  codec_tag_string: string;
+  codec_tag: string;
+
+  id?: string;
+  r_frame_rate?: string;
+  avg_frame_rate?: string;
+  time_base?: string;
+  start_pts?: string;
+  start_time?: string;
+  duration_ts?: string;
+  duration?: string;
+  bit_rate?: string;
+  max_bit_rate?: string;
+  bits_per_raw_sample?: number;
+  nb_frames?: number;
+  nb_read_frames?: number;
+  nb_read_packets?: number;
+  disposition?: RawProbeDisposition;
+
+  sample_fmt?: string;
+  sample_rate?: string;
+  channels?: number;
+
+  channel_layout?: string;
+  bits_per_sample?: number;
+  tags?: Record<string, string>;
+} | {
+  [key: string]: any;
+
+  index: number;
+  codec_name?: string;
+  codec_long_name?: string;
+  profile?: string;
+  codec_type: 'subtitle';
+  codec_time_base?: string;
+  codec_tag_string: string;
+  codec_tag: string;
+
+  id?: string;
+  r_frame_rate?: string;
+  avg_frame_rate?: string;
+  time_base?: string;
+  start_pts?: string;
+  start_time?: string;
+  duration_ts?: string;
+  duration?: string;
+  bit_rate?: string;
+  max_bit_rate?: string;
+  bits_per_raw_sample?: number;
+  nb_frames?: number;
+  nb_read_frames?: number;
+  nb_read_packets?: number;
+  disposition?: RawProbeDisposition;
+  width?: number;
+  height?: number;
+  tags?: Record<string, string>;
+} | {
+  [key: string]: any;
+
+  index: number;
+  codec_name?: string;
+  codec_long_name?: string;
+  profile?: string;
+  codec_type?: string;
+  codec_time_base?: string;
+  codec_tag_string: string;
+  codec_tag: string;
+
+  id?: string;
+  r_frame_rate?: string;
+  avg_frame_rate?: string;
+  time_base?: string;
+  start_pts?: string;
+  start_time?: string;
+  duration_ts?: string;
+  duration?: string;
+  bit_rate?: string;
+  max_bit_rate?: string;
+  bits_per_raw_sample?: number;
+  nb_frames?: number;
+  nb_read_frames?: number;
+  nb_read_packets?: number;
+  disposition?: RawProbeDisposition;
+  tags?: Record<string, string>;
+};
+
+/** @alpha */
+export interface RawProbeDisposition {
+  default: number;
+  dub: number;
+  original: number;
+  comment: number;
+  lyrics: number;
+  karaoke: number;
+  forced: number;
+  hearing_impaired: number;
+  visual_impaired: number;
+  clean_effects: number;
+  attached_pic: number;
+  timed_thumbnails: number;
+}
+
+// https://github.com/FFmpeg/FFmpeg/blob/9d8f9b2e4094ae6b07a9f23ae044b802722b3b4e/fftools/ffprobe.c#L2782
+/** @alpha */
+export interface RawProbeChapter {
+  id: number;
+  time_base: string;
+  start: number;
+  start_time: string;
+  end: number;
+  end_time: string;
+  tags?: Record<string, string>;
+}
+
+/* eslint-enable camelcase */
+
+/**
+ * **UNSTABLE**: `ProbeResult` is intended to have a simple API but it is still very unfinished, for
+ * the time being using `.unwrap()` is necessary to retrieve any useful information from `probe()`.
+ *
+ * @alpha
+ */
+export interface ProbeResult {
+  format?: Demuxer | Format | (string & {});
+  formatName?: string;
+
+  start: number;
+  duration: number;
+
+  bitrate?: number;
+  score: number;
+
+  tags?: Map<string, string>;
+
+  unwrap(): RawProbeResult;
+}
+
+/** @alpha */
+export interface ProbeOptions {
+  /**
+   * Specify the number of bytes to probe, if not given it will not be specified in the command-line
+   * arguments.
+   */
+  probeSize?: number;
+  /**
+   * Specify the number of milliseconds to analyze, defaults to `5000`.
+   */
+  analyzeDuration?: number;
+  /**
+   * Path to the `ffprobe` executable.
+   */
+  ffprobePath?: string;
+  /**
+   * Specify the input format of the media to probe.
+   */
+  format?: Demuxer | Format | (string & {});
+  /**
+   * Add command line arguments to ffprobe, `args` is appended **after** other
+   * arguments, but **before** source.
+   */
+  args?: string[];
+  /**
+   * Add custom options that will be used to spawn the process.
+   * {@link https://nodejs.org/docs/latest-v12.x/api/child_process.html#child_process_child_process_spawn_command_args_options}
+   * @example
+   * ```ts
+   * const info = await probe('video.mkv', {
+   *   spawnOptions: {
+   *     timeout: 5000
+   *   }
+   * });
+   * ```
+   */
+  spawnOptions?: SpawnOptions;
 }
