+++ test/process_test.ts
@@ -1,4 +1,3 @@
-import { FFmpegError } from '../src/errors';
 import { ffmpeg } from '../src/command';
 import { spawn } from '../src/process';
 import { isWin32 } from '../src/utils';
@@ -142,9 +141,8 @@ describe('process', function () {
         cmd.output()
           .args('-c', 'copy', '-f', 'matroska');
         const process = await cmd.spawn();
-        process.complete().then(console.log, console.log);
-        await expect(process.complete()).rejects.toThrow(FFmpegError);
-        await expect(process.complete()).rejects.toThrow(FFmpegError);
+        await expect(process.complete()).rejects.toThrow();
+        await expect(process.complete()).rejects.toThrow();
       });
       it('should reject on killed process', async function () {
         const cmd = ffmpeg();
@@ -153,8 +151,8 @@ describe('process', function () {
           .args('-c', 'copy', '-f', 'matroska');
         const process = await cmd.spawn();
         process.unwrap().kill();
-        await expect(process.complete()).rejects.toThrow(FFmpegError);
-        await expect(process.complete()).rejects.toThrow(FFmpegError);
+        await expect(process.complete()).rejects.toThrow();
+        await expect(process.complete()).rejects.toThrow();
       });
       it('should reject on errored process', async function () {
         const cmd = ffmpeg();
@@ -163,7 +161,6 @@ describe('process', function () {
         const process = await cmd.spawn({
           ffmpegPath: './my_invalid_ffmpeg'
         });
-        process.complete().catch(console.log);
         await expect(process.complete()).rejects.toThrow();
         await expect(process.complete()).rejects.toThrow();
       });
